+++ utils/zerovec/src/lib.rs
@@ -97,6 +97,7 @@
 
 extern crate alloc;
 
+mod error;
 pub mod map;
 #[cfg(test)]
 pub mod samples;
@@ -107,7 +108,7 @@ pub mod zerovec;
 #[cfg(feature = "yoke")]
 mod yoke_impls;
 
+pub use crate::error::ZeroVecError;
 pub use crate::map::ZeroMap;
-pub use crate::ule::ZeroVecError;
 pub use crate::varzerovec::VarZeroVec;
 pub use crate::zerovec::ZeroVec;
